ai-services/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
transformers==4.36.0
torch==2.1.0
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
